+++ src/address.rs
@@ -35,10 +35,14 @@ impl Address {
         self.d
     }
 
+    /// Returns the diversified base point for this address.
+    #[cfg_attr(feature = "unstable-voting-circuits", visibility::make(pub))]
     pub(crate) fn g_d(&self) -> NonIdentityPallasPoint {
         diversify_hash(self.d.as_array())
     }
 
+    /// Returns the diversified transmission key for this address.
+    #[cfg_attr(feature = "unstable-voting-circuits", visibility::make(pub))]
     pub(crate) fn pk_d(&self) -> &DiversifiedTransmissionKey {
         &self.pk_d
     }